Move hexToHue and hueToHex to color utils
